Final Review of Project

+++ creating-joy/src/main/java/com/technocrats/creatingjoy/service/AddressServiceImpl.java
@@ -1,15 +1,18 @@
 package com.technocrats.creatingjoy.service;
 
 import com.technocrats.creatingjoy.dao.AddressRepository;
+import com.technocrats.creatingjoy.dto.AddressDTO;
 import com.technocrats.creatingjoy.entity.Address;
+import com.technocrats.creatingjoy.objectmapper.AddressMapper;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class AddressServiceImpl implements AddressService {
 
 
@@ -17,46 +20,62 @@ public class AddressServiceImpl implements AddressService {
     private AddressRepository addressRepository;
 
     @Autowired
+    private AddressMapper addressMapper;
+
+    @Autowired
     public AddressServiceImpl(AddressRepository addressRepository){
         this.addressRepository=addressRepository;
     }
 
-    @Override
-    @Transactional
-    public List<Address> findAll() {
-        return addressRepository.findAll();
-    }
 
     @Override
     @Transactional
-    public Address findById(int id) {
+    public AddressDTO findById(int id) {
 
         Optional<Address> result=addressRepository.findById(id);
 
         if(result.isPresent()){
-            return result.get();
+            return addressMapper.convertToDto(result.get());
         }
         else{
             throw new RuntimeException("Did not found Address with id "+id);
         }
-
     }
 
     @Override
     @Transactional
-    public void save(Address address) {
+    public AddressDTO findByQueryId(int id) {
 
-        addressRepository.save(address);
+        Address result=addressRepository.findByQueryId(id);
+        if(result!=null){
+            return addressMapper.convertToDto(result);
+        }
+        else{
+            throw new RuntimeException("Did not found Query with id "+id);
+        }
+
     }
+
 
     @Override
     @Transactional
     public void deleteById(int id) {
 
-        addressRepository.deleteById(id);
+        if(id>0) {
+            addressRepository.deleteById(id);
+        }
 
     }
 
+    @Override
+    @Transactional
+    public void deleteByQueryId(int id) {
+        if(id>0) {
+
+            addressRepository.deleteByQueryId(id);
+        }
+
+    }
